requirements.txt
astroid==2.5.6
attrs==20.3.0
aws-cdk.assets==1.110.0
aws-cdk.aws-cloudwatch==1.110.0
aws-cdk.aws-ec2==1.110.0
aws-cdk.aws-events==1.110.0
aws-cdk.aws-iam==1.110.0
aws-cdk.aws-kms==1.110.0
aws-cdk.aws-logs==1.110.0
aws-cdk.aws-s3==1.110.0
aws-cdk.aws-s3-assets==1.110.0
aws-cdk.aws-ssm==1.110.0
aws-cdk.cloud-assembly-schema==1.110.0
aws-cdk.core==1.110.0
aws-cdk.cx-api==1.110.0
aws-cdk.region-info==1.110.0
cattrs==1.5.0
# Editable Git install with no remote (cdk-resources==0.0.1)
-e /home/martin/Source/cdk-resources/cdk_resources
constructs==3.3.90
isort==5.9.1
jsii==1.30.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
publication==0.0.3
pylint==2.8.3
python-dateutil==2.8.1
six==1.16.0
toml==0.10.2
typing-extensions==3.10.0.0
wrapt==1.12.1
